Don't persist globals as part of session state
We were not persistin them properly, because their env frame contains
interpreter values, which can't simply be serialized as JSON.  We have
the machinery to convert history values to JSON, because we have to
send them to the client.  But developing the corresponding machinery
for globals, which we don't want to force, will be a fair amount of work.
So just drop the globals from the persisted session state for now.